Finished Git integration. Front and back-end tweaks.

+++ assignaly_api/app/Http/Controllers/AssignmentController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\AssignmentStoreRequest;
 use App\Http\Requests\AssignmentUpdateRequest;
 use App\Http\Resources\AssignmentResource;
 use App\Models\Assignment;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Support\Facades\Auth;
@@ -14,8 +15,8 @@ class AssignmentController extends Controller
 {
     public function index(Request $request): AnonymousResourceCollection
     {
-        $assignments = Auth()->user()->ownedAssignments()
-            ->orderByDesc($request->input('orderBy') ?? 'created_at')
+        $assignments = Auth::user()->allAssignments()
+            ->orderByDesc('assignments.' . $request->input('orderBy') ?? 'created_at')
             ->paginate(perPage: $request->input('perPage') ?? 30);
 
         return AssignmentResource::collection($assignments);
@@ -23,16 +24,20 @@ class AssignmentController extends Controller
 
     public function latest(): AnonymousResourceCollection
     {
-        $assignments = Auth::user()->ownedAssignments()
-            ->orderByDesc('created_at')
+        $assignments = Auth::user()->allAssignments()
+            ->orderByDesc('assignments.created_at')
             ->take(5)
             ->get();
 
         return AssignmentResource::collection($assignments);
     }
 
-    public function store(AssignmentStoreRequest $request): AssignmentResource
+    public function store(AssignmentStoreRequest $request): AssignmentResource | JsonResponse
     {
+        if (Auth::user()->role->level < 2) {
+            return new JsonResponse(['error' => 'You\'re not authorized to create an assignment.'], 401);
+        }
+
         $validated = collect($request->validated())->except('users')->toArray();
         $users = $request->validated()['users'];
 
@@ -49,6 +54,10 @@ class AssignmentController extends Controller
 
     public function update(Assignment $assignment, AssignmentUpdateRequest $request): AssignmentResource
     {
+        if (Auth::user()->role->level < 2) {
+            return new JsonResponse(['error' => 'You\'re not authorized to update an assignment.'], 401);
+        }
+
         $validated = collect($request->validated())->except('users')->toArray();
 
         $assignment->update($validated);
@@ -63,6 +72,10 @@ class AssignmentController extends Controller
 
     public function destroy(Assignment $assignment): bool
     {
+        if (Auth::user()->role->level < 2) {
+            return new JsonResponse(['error' => 'You\'re not authorized to delete an assignment.'], 401);
+        }
+
         return $assignment->deleteOrFail();
     }
 }
